+++ src/controllers/Livro.controller.js
@@ -1,3 +1,4 @@
+import NaoEncontrado from "../Erros/NaoEncontrado.js";
 import livros from "../models/Livro.model.js";
 
 class livroController {
@@ -16,7 +17,11 @@ class livroController {
         try {
             const id = req.params.id;
             const livroEncontrado = await livros.findById(id);
-            res.status(200).json(livroEncontrado);
+            if (livroEncontrado !== null) {
+                res.status(200).send(livroEncontrado);
+            } else {
+                next(new NaoEncontrado("Id do livro não localizado."));
+            }
         } catch (erro) {
             next(erro);
         }
@@ -37,8 +42,13 @@ class livroController {
     static async alterarLivro(req, res, next) {
         try {
             const id = req.params.id;
+            const livroEncontrado = await livros.findById(id);
 
-            await livros.findByIdAndUpdate(id, { $set: req.body });
+            if (livroEncontrado !== null) {
+                await livros.findByIdAndUpdate(id, { $set: req.body });
+            } else {
+                next(new NaoEncontrado("Id do livro não localizado."));
+            }
 
             res.status(200).send({ message: "Livro atualizado com sucesso" });
         } catch (erro) {
@@ -49,8 +59,14 @@ class livroController {
     static async removerLivro(req, res, next) {
         try {
             const id = req.params.id;
-            await livros.findByIdAndDelete(id);
-            res.status(200).json({ message: "Livro excluido com sucesso!" });
+            const livroEncontrado = await livros.findById(id);
+
+            if (livroEncontrado !== null) {
+                await livros.findByIdAndDelete(id);
+                res.status(200).json({ message: "Livro excluido com sucesso!" });
+            } else {
+                next(new NaoEncontrado("Id do livro não localizado."));
+            }
         } catch (erro) {
             next(erro);
         }
